OOPified calcLinearIntercept.py
